+++ packages/vhd-lib/src/tests/utils.js
