+++ scripts/npmignore
@@ -30,6 +30,3 @@ __snapshots__/
 /docs/
 /src/
 /USAGE.md
-
-_testUtils.js
-_testUtils.js.map
